STT Trainer description
The NaomiSTTTrainer.py is not a plugin, it is a separate program
that uses plugins to train different speech to text engines to
your voice. I added a description of how that works.

Also fixed the password on the naobian page and tweaked some of
the information on the rasppi page.