Added RegExp
Added a small regexp to replace the three characters Phonetisaurus checks for ("_" - underscore, "}" - right curly bracket, and "|" - pipe) with spaces in the formatted cmu dictionary. This came up because I enabled the standard French model and phonetisaurus blew up because there was an underscore in the "a_contrario" entry. After applying this change, I was able to compile the French model correctly.