Aaron2020 Naomi-docs edits
This is my first commit of a total review of the documentation.
I'm just reading through the documentation and making changes when
I find things that are inaccurate. This commit includes fixing the
folder for Music Player Daemon plugin, fixing a few things in the
introduction, writing a new architecture image and updating the
readme for DeepSpeech, which was recently reimplemented with a
more streamlined setup and language model construction.